requirements.txt
nest_asyncio==1.5.8
flask==2.2.2
python-dotenv==1.0.0
openai==1.6.0
llama-index==0.9.19
llama-hub==0.0.61
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
whisper==1.1.10
pgvector==0.2.4
asyncpg==0.29.0
unstructured==0.11.2
cohere==4.37
Werkzeug==2.2.2
langchain==0.0.352
langchain_experimental==0.0.47
pika==1.3.2
google-cloud-storage==2.14.0
tiktoken==0.5.2
flask_swagger_ui==4.11.1
openai-whisper==20231117
ffmpeg-python==0.2.0
flask_cors==4.0.0
google-cloud-secret-manager==2.18.2
google-cloud==0.34.0
mysql-connector-python==8.4.0
pymssql==2.3.0
google-cloud-aiplatform==1.60.0